+++ js/jquery.terminal-1.4.3.js
@@ -31,7 +31,7 @@
  * Copyright (c) 2007-2013 Alexandru Marasteanu <hello at alexei dot ro>
  * licensed under 3 clause BSD license
  *
- * Date: Sun, 02 Jul 2017 17:59:17 +0000
+ * Date: Sun, 09 Jul 2017 12:23:42 +0000
  */
 
 /* TODO:
@@ -1699,8 +1699,18 @@
         // ---------------------------------------------------------------------
         var draw_prompt = (function() {
             function set(prompt) {
-                prompt_node.html($.terminal.format($.terminal.encode(prompt)));
-                prompt_len = prompt_node.text().length;
+                var lines = $.terminal.split_equal($.terminal.encode(prompt));
+                var formatted = lines.slice(0, -1).map(function(line) {
+                    line = $.terminal.format(line);
+                    if (line.match(/class="/)) {
+                        return line.replace(/class="/, 'class="line ');
+                    } else {
+                        return line.replace(/^<span/, '<span class="line"');
+                    }
+                }).concat(lines.slice(-1).map($.terminal.format)).join('\n');
+                var width = self.width();
+                prompt_node.html(formatted).find('.line').width(width);
+                prompt_len = prompt_node.find('span:last').text().length;
             }
             return function() {
                 switch (typeof prompt) {
@@ -1927,6 +1937,7 @@
                     change_num_chars();
                 }
                 redraw();
+                draw_prompt();
                 return self;
             },
             enable: function() {